Release 1.0

+++ app/controllers/requests_controller.rb
@@ -1,21 +1,37 @@
+# Impressum
+# Copyright by Maria Edlinger, Jonathan Lex and Markus Wallner
 class RequestsController < ApplicationController
-  before_action :set_request, only: [:accept, :reject]
-  before_action :logged_in, only: [:accept, :reject]
+  before_action :set_request, only: [:accept, :reject, :remove_confirmation]
 
   def accept
+    unless is_owner(@challenge)
+      redirect_to show_owner_challenge_path(@challenge), warning: 'You must be the owner of this challenge.'
+      return
+    end
     if @request.update(confirmed: true)
-      redirect_to show_owner_challenge_path(@challenge), notice: 'Successfully accepted.'
+      redirect_to show_owner_challenge_path(@challenge), success: 'Successfully accepted.'
+      return
     else
-      redirect_to show_owner_challenge_path(@challenge), notice: 'NOT successfully accepted.'
+      redirect_to show_owner_challenge_path(@challenge), error: 'Oops, there was a problem with your acception. Please try again.'
+      return
     end
   end
 
   def reject
-    if @request.destroy
-      redirect_to show_owner_challenge_path(@challenge), notice: 'Successfully removed.'
-    else
-      redirect_to show_owner_challenge_path(@challenge), notice: 'NOT successfully removed.'
+    unless is_owner(@challenge)
+      redirect_to show_owner_challenge_path(@challenge), warning: 'You must be the owner of this challenge.'
+      return
     end
+    if @request.destroy
+      redirect_to show_owner_challenge_path(@challenge), success: 'Successfully removed.'
+      return
+    else
+      redirect_to show_owner_challenge_path(@challenge), error: 'Oops, there was a problem with your rejection. Please try again.'
+      return
+    end
+  end
+
+  def remove_confirmation
   end
 
   private
@@ -23,11 +39,5 @@ class RequestsController < ApplicationController
   def set_request
     @request = Request.find(params[:id])
     @challenge = @request.challenge
-  end
-
-  def logged_in
-    unless current_user
-      redirect_to signup_path, notice: 'You must be logged in.'
-    end
   end
 end
